Implementing VIEW PRINT parsing
