+++ src/linter/mod.rs
@@ -1,6 +1,6 @@
 mod const_value_resolver;
 mod converter;
-mod post_linter;
+pub mod post_linter;
 mod pre_linter;
 mod type_resolver;
 mod type_resolver_impl;
